fix gateway proto (#29)
* fix gateway proto

* json response balance

* json header

* json header

* calculation fix

* user order number registration fix

* fix order registration time

+++ internal/gophermart/domain/usecase/listUserWithdraw.go
@@ -19,7 +19,7 @@ type (
 
 	ListUserWithdrawalsOutputDTO struct {
 		ProcessedAt string             `json:"processed_at"` // nolint:tagliatelle // external requirements
-		Order       int                `json:"order"`
+		Order       int64              `json:"order"`
 		Sum         sharedkernel.Money `json:"sum"`
 	}
